feat(console): add `isCloud` constant
