+++ packages/console/src/cloud/components/CardSelector/MultiCardSelector/index.module.scss
