+++ packages/console/src/cloud/components/ProgressBar/index.module.scss
