+++ packages/console/src/cloud/pages/Congrats/index.module.scss
