+++ packages/console/src/cloud/scss/layout.module.scss
