+++ packages/console/src/consts/cloud.ts
@@ -1,0 +1,3 @@
+import { yes } from '@silverhand/essentials';
+
+export const isCloud = yes(process.env.IS_CLOUD);
